Move query-serialization
